requirements.txt
appdirs==1.4.4
asgiref
beautifulsoup4
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django
django-bootstrap4
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-extensions==2.2.5
django-js-asset==1.2.2
django-profanity-filter==0.2.1
django-widget-tweaks==1.4.1
filelock==3.0.12
greenlet==0.4.15
idna==2.8
Markdown==2.6.9
MarkupPy==1.14
model-mommy==1.3.2
msgpack==1.0.0
oauthlib==3.1.0
pep8==1.7.1
Pillow==7.1.1
psycopg2-binary
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.1.2
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.0
urllib3==1.25.2
virtualenv==20.0.21
